refactor: separate messages to pin from suggestions
