+++ mp2i/cogs/suggestions.py
@@ -16,8 +16,6 @@ class Suggestion(Cog):
     Offers commands to allow members to propose suggestions and interact with them
     """
 
-    MINIMUM_PINS = 5
-
     def __init__(self, bot):
         self.bot = bot
 
@@ -116,40 +114,6 @@ class Suggestion(Cog):
 
         await channel.send(file=file, embed=embed)
         await suggestion.delete()
-
-    @Cog.listener("on_raw_reaction_add")
-    @guild_only()
-    async def add_pin(self, payload) -> None:
-        """
-        Add a pin to a message and send it to website channel when
-        it reach the required number of pins reactions.
-        """
-        if str(payload.emoji) != "📌":
-            return
-
-        channel = self.bot.get_channel(payload.channel_id)
-        message = await channel.fetch_message(payload.message_id)
-        pins = discord.utils.get(message.reactions, emoji="📌", me=False)
-        if pins is None or pins.count < self.MINIMUM_PINS:
-            return
-
-        author = message.author
-        embed = discord.Embed(
-            colour=0x00FF00,
-            title="Message épinglé",
-            description="Un message a été retenu par la communauté, vous pouvez "
-            "probablement l'ajouter dans la [FAQ](https://prepas-mp2i.fr/faq/).",
-            timestamp=datetime.now(),
-        )
-        embed.add_field(name="Lien du message", value=message.jump_url)
-        embed.set_author(name=author.name, icon_url=author.avatar.url)
-        embed.set_footer(text=self.bot.user.name)
-        website_chan = self.bot.get_channel(
-            GuildWrapper(channel.guild).config.channels.website
-        )
-        await website_chan.send(embed=embed)
-        # Pour ne pas envoyer le message plusieurs fois
-        await message.add_reaction("📌")
     
     @hybrid_command(name="suggestions")
     @guild_only()
